# Faster collection truncating with deleteByQuery