--- a/doc/2/api/controllers/collection/truncate/index.md
+++ b/doc/2/api/controllers/collection/truncate/index.md
@@ -19,7 +19,7 @@ Documents removed that way do not trigger real-time notifications.
 ### HTTP
 
 ```http
-URL: http://kuzzle:7512/<index>/<collection>/_truncate[?strategy=<documents|collection>]
+URL: http://kuzzle:7512/<index>/<collection>/_truncate
 Method: DELETE
 ```
 
@@ -30,10 +30,7 @@ Method: DELETE
   "index": "<index>",
   "collection": "<collection>",
   "controller": "collection",
-  "action": "truncate",
-
-  // optional
-  "strategy": "<documents|collection>"
+  "action": "truncate"
 }
 ```
 
@@ -43,20 +40,8 @@ Method: DELETE
 
 - `collection`: collection name
 - `index`: index name
-- `strategy`: truncate strategy <SinceBadge version="auto-version" />
 
 ---
-
-### strategy
-
-<SinceBadge version="auto-version" />
-
-Two strategies are available when truncating a collection:
-  - `collection`: delete and re-create the underlaying collection
-  - `documents`: search and deletes every documents in the collection
-
-Both methods offer similar performances when truncating a collection with < 10 million documents.    
-However when truncating many collection, it's faster to use the `documents` strategy because the storage engine is able to perform those actions simultaneously.
 
 ## Response