--- a/lib/api/controllers/collectionController.js
+++ b/lib/api/controllers/collectionController.js
@@ -271,13 +271,29 @@ class CollectionController extends NativeController {
   /**
    * Reset a collection by removing all documents while keeping the existing mapping.
    *
-   * @param {Request} request
+   * @param {KuzzleRequest} request
    * @returns {Promise.<Object>}
    */
   async truncate (request) {
     const { index, collection } = request.getIndexAndCollection();
+    const strategy = request.getString('strategy', 'documents')
 
-    await this.ask('core:storage:public:collection:truncate', index, collection);
+    if (! ['collection', 'documents'].includes(strategy)) {
+      throw kerror.get('api', 'assert', 'invalid_argument', '"collection", "documents"');
+    }
+
+    if (strategy === 'collection') {
+      await this.ask('core:storage:public:collection:truncate', index, collection);
+    }
+    else {
+      await this.ask('core:storage:public:collection:refresh', index, collection);
+      await this.ask(
+        'core:storage:public:document:deleteByQuery',
+        index,
+        collection,
+        {},
+        { refresh: 'wait_for', fetch: false });
+    }
 
     return {
       acknowledged: true
@@ -294,7 +310,7 @@ class CollectionController extends NativeController {
     const size = request.getInteger('size', 0);
     const type = request.getString('type', 'all');
 
-    if (['all', 'stored', 'realtime'].indexOf(type) === -1) {
+    if (! ['all', 'stored', 'realtime'].includes(type)) {
       throw kerror.get('api', 'assert', 'invalid_argument', '"all", "stored", "realtime"');
     }