--- a/lib/api/controllers/collectionController.js
+++ b/lib/api/controllers/collectionController.js
@@ -269,31 +269,21 @@ class CollectionController extends NativeController {
   }
 
   /**
-   * Reset a collection by removing all documents while keeping the existing mapping.
+   * Reset a collection by removing all documents
    *
    * @param {KuzzleRequest} request
    * @returns {Promise.<Object>}
    */
   async truncate (request) {
     const { index, collection } = request.getIndexAndCollection();
-    const strategy = request.getString('strategy', 'documents')
 
-    if (! ['collection', 'documents'].includes(strategy)) {
-      throw kerror.get('api', 'assert', 'invalid_argument', '"collection", "documents"');
-    }
-
-    if (strategy === 'collection') {
-      await this.ask('core:storage:public:collection:truncate', index, collection);
-    }
-    else {
-      await this.ask('core:storage:public:collection:refresh', index, collection);
-      await this.ask(
-        'core:storage:public:document:deleteByQuery',
-        index,
-        collection,
-        {},
-        { refresh: 'wait_for', fetch: false });
-    }
+    await this.ask('core:storage:public:collection:refresh', index, collection);
+    await this.ask(
+      'core:storage:public:document:deleteByQuery',
+      index,
+      collection,
+      {},
+      { fetch: false, refresh: 'wait_for' });
 
     return {
       acknowledged: true