--- a/test/api/controllers/collectionController.test.js
+++ b/test/api/controllers/collectionController.test.js
@@ -119,13 +119,30 @@ describe('Test: collection controller', () => {
       const response = await collectionController.truncate(request);
 
       should(kuzzle.ask).be.calledWith(
-        'core:storage:public:collection:truncate',
+        'core:storage:public:collection:refresh',
         index,
         collection);
+      should(kuzzle.ask).be.calledWith(
+        'core:storage:public:document:deleteByQuery',
+        index,
+        collection,
+        {},
+        { refresh: 'wait_for', fetch: false });
 
       should(response).match({
         acknowledged: true
       });
+    });
+
+    it('should allows to use another truncate strategy', async () => {
+      request.input.args.strategy = 'collection';
+
+      await collectionController.truncate(request);
+
+      should(kuzzle.ask).be.calledWith(
+        'core:storage:public:collection:truncate',
+        index,
+        collection);
     });
   });