## What does this PR do ?

When truncating a collection, Kuzzle delete and then re-create the underlying Elasticsearch indice.   
However indice creation is not parallelizable inside an ES cluster so if we truncate many collections it will be quite slow when the deleteByQuery action is completely parallel. 

This PR change `collection:truncate` to use `deleteByQuery` which is almost always faster. 

```
// 2 000 000 documents
Big collection: delete + create collection: 192.851ms
Big collection: delete documents: 347.221ms

// 10 000 documents
Medium collection: delete + create collection: 186.285ms
Medium collection: delete documents: 90.816ms

// 1000 documents
Small collection: delete + create collection: 139.613ms
Small collection: delete documents: 79.563ms

Truncate 10 collections in parallel: documents: 64.560ms
Truncate 10 collections in parallel: collection: 1263.654ms
```