@Aschen 2021-07-20
> Ok for me. But have you tested the solution of deleting and recreating the index? Maybe it's even faster

That's the whole purpose of this PR :sweat_smile: Before we were deleting and then creating the ES index but it was slower than using `deleteByQuery`